+++ model/src/main/scala/com/codefactory/ecommerce/tableModel/Bank.scala
@@ -1,3 +1,0 @@
-package com.codefactory.ecommerce.tableModel
-
-final case class Bank(id: Int, balance: Int)