+++ model/src/main/scala/com/codefactory/ecommerce/tableModel/Order.scala
@@ -1,3 +1,3 @@
 package com.codefactory.ecommerce.tableModel
 
-final case class Order(id: Int, userID: Int, status: String)
+final case class Order(id: Int, userID: Int, status: String, orderCartID: Int)