+++ model/src/main/scala/com/codefactory/ecommerce/tableModel/Product.scala
@@ -2,7 +2,8 @@ package com.codefactory.ecommerce.tableModel
 
 final case class Product(id: Int,
                          name: String,
-                         price: Int, //Int pentru pret e idee rea sau eu nu am vazut note in privinta preciziei.
-                         typeID: Int,
+                         price: Float, //Changed to float for more precision
+                         productType: String,
                          quantity: Int,
-                         isDeleted: Boolean)
+                         isDeleted: Boolean,
+                         cartProductID: Int)