+++ model/src/main/scala/com/codefactory/ecommerce/tableModel/Type.scala
@@ -1,3 +1,0 @@
-package com.codefactory.ecommerce.tableModel
-
-final case class Type(id: Int, tname: String)