+++ model/src/main/scala/com/codefactory/ecommerce/tableModel/User.scala
@@ -1,3 +1,3 @@
 package com.codefactory.ecommerce.tableModel
 
-final case class User(id: Int, name: String, email: String, bankID: Int)
+final case class User(id: Int, name: String, email: String, bankBalance: Float)