+++ persistence/src/main/scala/com/codefactory/ecommerce/tableQuerryVariable/QueryVariable.scala
@@ -1,26 +1,15 @@
 package com.codefactory.ecommerce.tableQuerryVariable
 
-import com.codefactory.ecommerce.tableDef.{
-  BankT,
-  CartT,
-  OrderT,
-  ProductT,
-  TypeT,
-  UserT
-}
+import com.codefactory.ecommerce.tableDef.{CartT, OrderT, ProductT, UserT}
 import slick.lifted.TableQuery
 
 import slick.jdbc.MySQLProfile.api._
 
 trait QueryVariable {
 
-//  val db = Database.forConfig("mydb")
-
-  final val banks = TableQuery[BankT]
   final val carts = TableQuery[CartT]
   final val orders = TableQuery[OrderT]
   final val products = TableQuery[ProductT]
-  final val types = TableQuery[TypeT]
   final val users = TableQuery[UserT]
 
 }