+++ src/main/resources/db/migration/V2__INSERT INTO TABLES.sql
@@ -1,10 +1,3 @@
-INSERT INTO USERS VALUES (1,'Ciobanu Alin','alin.m.ciobanu@gmail.com',1);
-INSERT INTO USERS VALUES (2,'Testing Migration','testingmigration@gmail.com',2);
-INSERT INTO BANK VALUES (1,1000);
-INSERT INTO BANK VALUES (2,2000);
-INSERT INTO PRODUCTS VALUES(1,'cana',20,1,100,0);
-INSERT INTO PRODUCTS VALUES(2,'vaza',80,2,20,0);
-INSERT INTO PRODUCTS VALUES(3,'pian',5000,3,3,0);
-INSERT INTO TYPES VALUES(1,'vesela');
-INSERT INTO TYPES VALUES(2,'decoratiuni');
-INSERT INTO TYPES VALUES(3,'muzica');
+INSERT INTO USERS VALUES (1,'Ciobanu Alin','alin.m.ciobanu@gmail.com',100);
+INSERT INTO USERS VALUES (2,'Testing Migration','testingmigration@gmail.com',200);
+